+++ docs/src/news.md
@@ -1,5 +1,7 @@
 ### NEWS for package **Jchemo**
 
+### *Version 0.1.18*
+
 ### *Version 0.1.17*
 
 - News